Remove usage of `hierplace`
Some basic profiling with python's `cProfile` shows that `hierplace` is
significant bottleneck of this script (at least when using KiCad7).
See below measurements:

```
225980566 function calls (218237324 primitive calls) in 73.127 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000   73.127   73.127 kinet2pcb.py:294(main)
        1    0.003    0.003   73.125   73.125 kinet2pcb.py:187(kinet2pcb)
        1    0.000    0.000   63.472   63.472 HierPlace.py:274(hier_place)
```

Out of total 73s running `main`, `hier_place` took 63s!
Because `hierplace` is not required for my use-case (maintaining fork
to be used in [1]) decided to just remove it.

[1] https://github.com/adamws/keyboard-tools